+++ pluggable-architecture/git-stats/git_stats/download.py
@@ -1,6 +1,6 @@
-from typing import NamedTuple
 from urllib.parse import urlparse
 
+from .dto import RepoDetails, RepoStatistics
 from .plugins.bitbucket_plugin import BitBucketPlugin
 from .plugins.github_plugin import GitHubPlugin
 from .plugins.gitlab_plugin import GitLabPlugin
@@ -8,15 +8,7 @@ from .plugins.gitlab_plugin import GitLabPlugin
 plugins = [BitBucketPlugin, GitHubPlugin, GitLabPlugin]
 
 
-class RepoDetails(NamedTuple):
-    organization: str
-    repo: str
-
-    def __str__(self):
-        return self.organization + "/" + self.repo
-
-
-class GitFetcher:
+class GitApiClient:
     def __init__(self, url):
         domain, self.repo = self._parse_url(url)
         for plugin in plugins:
@@ -31,5 +23,5 @@ class GitFetcher:
         parts = url_parts.path.split("/")
         return url_parts.netloc, RepoDetails(organization=parts[1], repo=parts[2])
 
-    def get_stats(self):
+    def get_stats(self) -> RepoStatistics:
         return self.plugin.repo_stats()
