+++ pluggable-architecture/git-stats/tests/test_download.py
@@ -1,10 +1,10 @@
-from git_stats.download import GitFetcher
+from git_stats.download import GitApiClient
 
 
 def test_url_parser():
     url = "https://gitlab.com/pycqa/flake8"
 
-    client = GitFetcher(url)
+    client = GitApiClient(url)
 
     repo_details = client.repo
     assert repo_details.organization == "pycqa"
